+++ services/dask-sidecar/src/simcore_service_dask_sidecar/rabbitmq_worker_plugin.py
@@ -24,7 +24,7 @@ _RABBITMQ_CONFIGURATION_ERROR: Final[str] = (
 class RabbitMQPlugin(distributed.WorkerPlugin):
     """Dask Worker Plugin for RabbitMQ integration"""
 
-    name = "rabbitmq_plugin"
+    name = "rabbitmq_worker_plugin"
     _main_thread_loop: AbstractEventLoop | None = None
     _client: RabbitMQClient | None = None
     _settings: RabbitSettings | None = None
