+++ services/dask-sidecar/tests/unit/test_worker.py
@@ -113,7 +113,8 @@ def dask_subsystem_mock(
     )
     # mock dask rabbitmq plugin
     mock_dask_rabbitmq_plugin = mocker.patch(
-        "simcore_service_dask_sidecar.rabbitmq_plugin.RabbitMQPlugin", autospec=True
+        "simcore_service_dask_sidecar.rabbitmq_worker_plugin.RabbitMQPlugin",
+        autospec=True,
     )
     mock_rabbitmq_client = create_rabbitmq_client("pytest_dask_sidecar_logs_publisher")
     mock_dask_rabbitmq_plugin.get_client.return_value = mock_rabbitmq_client
